pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "llm-quanta"
version = "0.1.0"
description = "Unified LLM quantization CLI with automated benchmarking and comparison"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [{ name = "Your Name", email = "you@example.com" }]
keywords = ["llm", "quantization", "gptq", "awq", "gguf", "bitsandbytes", "benchmark"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "torch>=2.0.0",
    "transformers>=4.36.0",
    "accelerate>=0.25.0",
    "safetensors>=0.4.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "datasets>=2.14.0",
    "ninja>=1.11.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "tqdm>=4.65.0",
]

[project.optional-dependencies]
gptq = ["auto-gptq>=0.7.0"]
awq = ["autoawq>=0.2.0"]
bitsandbytes = ["bitsandbytes>=0.43.0"]
gguf = ["llama-cpp-python>=0.2.0"]
all = [
    "auto-gptq>=0.7.0",
    "autoawq>=0.2.0",
    "bitsandbytes>=0.43.0",
    "llama-cpp-python>=0.2.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
llm-quanta = "llm_quanta.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/llm-quanta"
Documentation = "https://github.com/yourusername/llm-quanta#readme"
Repository = "https://github.com/yourusername/llm-quanta"

[tool.hatch.build.targets.wheel]
packages = ["src/llm_quanta"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["tests"]
